updateReact and updateAPI

+++ Api/VidShare/VidShare.API/Controllers/UserController.cs
@@ -1,5 +1,9 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using System.Diagnostics.CodeAnalysis;
+using VidShare.API.Models;
+using VidShare.Core.DTOs;
 using VidShare.Core.Models;
 using VidShare.Core.Services;
 using VidShare.Service;
@@ -10,20 +14,24 @@ namespace VidShare.API.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
+    [Authorize]
     public class UserController : ControllerBase
     {
         private readonly IUserService _userService;
-        public UserController(IUserService userService)
+        private readonly IMapper _mapper;
+        public UserController(IUserService userService,IMapper mapper)
         {
             _userService = userService;
+            _mapper=mapper;
         }
 
         // GET: api/<UsersController>
         [HttpGet]
-        public ActionResult Get()
+        public async Task<ActionResult> Get()
         {
-            var users= _userService.GetAll();
-            return Ok(users);
+            var list=await _userService.GetAllAsync();
+            var listDTO=_mapper.Map<IEnumerable<UserDTO>>(list);
+            return Ok(listDTO);
         }
 
         // GET api/<UsersController>/5
@@ -31,29 +39,45 @@ namespace VidShare.API.Controllers
         public ActionResult Get(int id)
         {
             var user= _userService.GetById(id);
-            return Ok(user);
+            var userDTO= _mapper.Map<UserDTO>(user);   
+            return Ok(userDTO);
         }
 
         // POST api/<UsersController>
         [HttpPost]
-        public ActionResult Post([FromBody] User user)
+        public async Task<ActionResult> Post([FromBody] UserPostModel user)
         {
-            var newUser = _userService.Add(user);
-            return Ok(newUser);
+            var newUser = new User()
+            {
+                  Name = user.Name,
+                  Email = user.Email,
+                  Password = user.Password,
+                  Role = user.Role, 
+            };
+            var UserNew=await _userService.AddAsync(newUser);
+            return Ok(UserNew);
         }
 
         // PUT api/<UsersController>/5
         [HttpPut("{id}")]
-        public ActionResult Put(int id, [FromBody] User user)
+        public ActionResult Put(int id, [FromBody] UserPostModel user)
         {
-            var updateUser = _userService.Update(user);
-            return Ok(updateUser);
+            var updateUser = new User()
+            {
+                Name = user.Name,
+                Email = user.Email,
+                Password = user.Password,
+                Role = user.Role,
+            };
+            return Ok(_userService.Update(updateUser));
         }
 
         // DELETE api/<UsersController>/5
         [HttpDelete("{id}")]
         public ActionResult Delete(int id)
         {
+            var user = _userService.GetById(id);
+            if(user is null) { return NotFound(); }
             _userService.Delete(id);
             return Ok("user deleted successfully");    
         }
